refactor(utils): reformatted the stack trace to a validity tree using classes and interfaces
The StackTrace was replaced with a ValidityTree so that documentation of how validateTSRObj reports
invalid fields can be more comprehensive. Additionally, encapsulating the validity object as classes
and interfaces, related through inheritence, better conveys the concept of each type of node in the
validity tree by using the ubiquitous and commonly understood framework of OOP.  Lastly, the data
type returned by validateTSRObj is a tree - so calling it a stack trace is misleading.